+++ src/pages/UserPage/ProfileBookList/tests/ProfileBookList.test.js
@@ -1,6 +1,6 @@
 import { shallow } from 'enzyme';
-import ProfileBookList from '../components/ProfileBookList';
-import ProfileBookListItem from '../components/ProfileBookListItem';
+import ProfileBookList from '../ProfileBookList';
+import ProfileBookListItem from '../ProfileBookListItem/ProfileBookListItem';
 import { Typography } from '@material-ui/core';
 
 const testProps = [
