+++ src/pages/UserPage/index.tsx
@@ -1,13 +1,6 @@
-import ProfileData from './components/ProfileData';
-import ProfileHeader from './components/ProfileHeader';
-import ProfileBookList from './components/ProfileBookList';
-import ProfileBookListItem from './components/ProfileBookListItem';
-import ProfileImage from './components/ProfileImage';
+import ProfileData from './ProfileData/ProfileData';
+import ProfileHeader from './ProfileHeader/ProfileHeader';
+import ProfileBookList from './ProfileBookList/ProfileBookList';
+import ProfileImage from './ProfileImage/ProfileImage';
 
-export {
-  ProfileData,
-  ProfileHeader,
-  ProfileImage,
-  ProfileBookList,
-  ProfileBookListItem,
-};
+export { ProfileData, ProfileHeader, ProfileImage, ProfileBookList };
